Improved performance by a factor of ~10
PhoneRegex now uses Lazy<Regex> and the RegexCache now uses a ConcurrentDictionary internally.

I have sacrificed the LRU cache, which is why the TestRegexInsertion currently fails. This means Regexes are currently cached indefinitely. I wonder in how many real-world scenario's caching a few hundred(?) regexes would be an actual problem. Unfortunately I have no numbers on the use of this class nor any idea about the environments this class is used in. But the added complexity of the LRU (and making it thread-safe) seems not worth it to me. Especially if this easy optimization alone results in a 10x improvement.

My numbers for running all unittests went from ~18seconds to ~1.8 seconds on my 9 year old workstation. Will discuss further in #136